Replace No Clip with Collision Mode
Adds a new "opaque" collision mode, fixing #55

Replaces No Clip being enabled with the "none" mode and disabled with the "all" mode.

Uses `material.blocksLight()` to decide which blocks are opaque, if
this leads to issues we could manually curate a transparent block list
or find an alternative method of deciding.